style(types): move type of object to types dir
